files to create exe
added files to create the executable, as far as i tested works with python 3.10.X, and doesn't work with >3.11 or >3.12. You can change the PYTHON_PATH variable inside the .bat to the python.exe path in your computer in case that one doesn't work.
You can also just download the exe .zip and avoid all that.